Feat: userController 정리
